Add volunteer names to session_detail
